+++ src/Sylius/Bundle/AddressingBundle/spec/Sylius/Bundle/AddressingBundle/Form/DataTransformer/ZoneToIdentifierTransformerSpec.php
@@ -5,8 +5,6 @@ namespace spec\Sylius\Bundle\AddressingBundle\Form\DataTransformer;
 use PHPSpec2\ObjectBehavior;
 
 /**
- * Zone to identifier transformer spec.
- *
  * @author Paweł Jędrzejewski <pjedrzejewski@diweb.pl>
  */
 class ZoneToIdentifierTransformer extends ObjectBehavior
@@ -19,12 +17,12 @@ class ZoneToIdentifierTransformer extends ObjectBehavior
         $this->beConstructedWith($zoneRepository, 'name');
     }
 
-    function it_should_be_initializable()
+    function it_is_initializable()
     {
         $this->shouldHaveType('Sylius\Bundle\AddressingBundle\Form\DataTransformer\ZoneToIdentifierTransformer');
     }
 
-    function it_should_return_empty_string_if_null_transormed()
+    function it_returns_empty_string_if_null_transormed()
     {
         $this->transform(null)->shouldReturn('');
     }
@@ -49,12 +47,12 @@ class ZoneToIdentifierTransformer extends ObjectBehavior
         $this->transform($zone)->shouldReturn('EU');
     }
 
-    function it_should_return_null_if_empty_string_reverse_transformed()
+    function it_returns_null_if_empty_string_reverse_transformed()
     {
         $this->reverseTransform('')->shouldReturn(null);
     }
 
-    function it_should_return_null_if_zone_not_found_on_reverse_transform($zoneRepository)
+    function it_returns_null_if_zone_not_found_on_reverse_transform($zoneRepository)
     {
         $zoneRepository
             ->findOneBy(array('name' => 'EU'))