+++ src/Sylius/Bundle/AddressingBundle/CHANGELOG.md
@@ -1,6 +1,15 @@
 CHANGELOG
 =========
 
+### v0.2.0
+
+* Symfony 2.3 upgrade.
+* Removal of ``Entity`` namespace, extend ``Model`` instead.
+
+### v0.1.0
+
+* Initial dev release, last compatible version with Symfony 2.3.
+
 ### 11-02-2013
 
 * Bump the Symfony requirement to version 2.2.